Make use of AC_CONFIG_MACRO_DIRS
Make use of AC_CONFIG_MACRO_DIRS which was introduced in
autoconf 2.70 with automake 1.13.2 providing a fallback
solution for autoconf before 2.70.

We update the automake requirement not to just 1.13.2 but
to 1.14 because many other packages I maintain require that
for good reasons.

This means we can remove the macro directories from the
ACLOCAL_AMFLAGS.